+++ legacy/src/api/nomenclatures.ts
@@ -1,15 +1,27 @@
-import getNomenclaturesContent from './data/get-nomenclatures.json';
+import { computeAuthorizationHeader, getBaseURI } from './utils';
 
-type NomenclaturesJSON = {
-  nomenclatures: { [key: string]: unknown };
+export type Nomenclature = {
+  id: string;
+  /**
+   * @deprecated (name always equals to id)
+   */
+  name?: string;
+  label: string;
+  /**
+   * can be a integer (from registry) or a string (version = name actually)
+   */
+  version: number | string;
+  urn?: string;
+  suggesterParameters: unknown; // suggester configuration (abstract in pogues)
 };
 
-const nomenclaturesJSON: NomenclaturesJSON = getNomenclaturesContent;
-
-export function getNomenclatures(): NomenclaturesJSON {
-  return nomenclaturesJSON;
-}
-
-export function getNomenclature(id: string): unknown {
-  return nomenclaturesJSON.nomenclatures[id];
+export async function getNomenclatures(token: string): Promise<Nomenclature[]> {
+  const url = `${getBaseURI()}/nomenclatures`;
+  const headers = new Headers();
+  headers.append('Accept', 'application/json');
+  headers.append('Authorization', computeAuthorizationHeader(token));
+  const nomenclatures = await fetch(url, { headers }).then(
+    (res) => res.json() as Promise<Nomenclature[]>,
+  );
+  return nomenclatures;
 }